fix: bumps and bootloader only support
* Move to mainline ATF
* Update u-boot to latest release
* Rebase kernel to rk3588-test branch for testing features
* Move debian mirror back to normal URL
* Add support to skip debootstrap, and generate only kernel + u-boot image.

+++ scripts/03_docker.sh
@@ -43,7 +43,10 @@ fi
 debug_msg "Docker: Generating Disk Images..."
 docker run --ulimit nofile=1024 --rm --privileged --cap-add=ALL -v "/dev:/dev:Z" -v "${root_path}:/repo:Z" -it ${docker_tag} /repo/scripts/docker/build_image.sh
 
-debug_msg "Docker: debootstraping..."
-docker run --ulimit nofile=1024 --rm --privileged --cap-add=ALL -v "/dev:/dev:Z" -v "${root_path}:/repo:Z" -it ${docker_tag} /repo/scripts/docker/run_debootstrap.sh
+# Only debootstrap on full build
+if [ -z "${BOOTLOADER_ONLY}" ]; then
+    debug_msg "Docker: debootstraping..."
+    docker run --ulimit nofile=1024 --rm --privileged --cap-add=ALL -v "/dev:/dev:Z" -v "${root_path}:/repo:Z" -it ${docker_tag} /repo/scripts/docker/run_debootstrap.sh
+fi
 
 debug_msg "Finished 03_docker.sh"
